requirements.txt
defusedxml==0.7.1
fonttools==4.57.0
fpdf2==2.8.2
pillow==11.2.1
